+++ src/wildboar/distance/_elastic.pyx
@@ -1134,6 +1134,94 @@ cdef double erp_distance(
     return cost_prev[y_length - 1]
 
 
+cdef double erp_subsequence_distance(
+    const double *S,
+    Py_ssize_t s_length,
+    const double *T,
+    Py_ssize_t t_length,
+    Py_ssize_t r,
+    double g,
+    double *gX,
+    double *gY,
+    double *cost,
+    double *cost_prev,
+    Py_ssize_t *index,
+) nogil:
+    cdef double dist = 0
+    cdef double min_dist = INFINITY
+
+    cdef Py_ssize_t i
+    for i in range(t_length - s_length + 1):
+        dist = erp_distance(
+            S,
+            s_length,
+            T + i,
+            s_length,
+            r, 
+            g,
+            gX,
+            gY,
+            cost, 
+            cost_prev, 
+        )
+
+        if dist < min_dist:
+            if index != NULL:
+                index[0] = i
+            min_dist = dist
+
+    return min_dist
+
+
+cdef Py_ssize_t erp_subsequence_matches(
+    const double *S,
+    Py_ssize_t s_length,
+    const double *T,
+    Py_ssize_t t_length,
+    Py_ssize_t r,
+    double g,
+    double *gX,
+    double *gY,
+    double *cost,
+    double *cost_prev,
+    double threshold,
+    double **distances,
+    Py_ssize_t **matches
+) nogil:
+    cdef double dist = 0
+    cdef Py_ssize_t capacity = 1
+    cdef Py_ssize_t tmp_capacity
+    cdef Py_ssize_t i
+    cdef Py_ssize_t n_matches = 0
+
+    matches[0] = <Py_ssize_t*> malloc(sizeof(Py_ssize_t) * capacity)
+    distances[0] = <double*> malloc(sizeof(double) * capacity)
+
+    for i in range(t_length - s_length + 1):
+        dist = erp_distance(
+            S,
+            s_length,
+            T + i,
+            s_length,
+            r, 
+            g,
+            gX,
+            gY,
+            cost, 
+            cost_prev, 
+        )
+
+        if dist <= threshold:
+            tmp_capacity = capacity
+            realloc_array(<void**> matches, n_matches, sizeof(Py_ssize_t), &tmp_capacity)
+            realloc_array(<void**> distances, n_matches, sizeof(double), &capacity)
+            matches[0][n_matches] = i
+            distances[0][n_matches] = dist
+            n_matches += 1
+
+    return n_matches
+
+
 cdef double edr_distance(
     const double *X,
     Py_ssize_t x_length,
@@ -2476,6 +2564,114 @@ cdef class MsmSubsequenceDistanceMeasure(SubsequenceDistanceMeasure):
             self.cost,
             self.cost_prev,
             self.cost_y,
+            threshold,
+            distances,
+            indicies,
+        )
+
+
+cdef class ErpSubsequenceDistanceMeasure(SubsequenceDistanceMeasure):
+
+    cdef double *cost
+    cdef double *cost_prev
+    cdef double *gX
+    cdef double *gY
+    cdef double r
+    cdef double g
+    
+    def __init__(self, double r=1.0, double g=0.0):
+        check_scalar(r, "r", float, min_val=0.0, max_val=1.0)
+        check_scalar(g, "g", float, min_val=0)
+        self.r = r
+        self.g = g
+
+    def __cinit__(self, *args, **kwargs):
+        self.cost = NULL
+        self.cost_prev = NULL
+        self.gX = NULL
+        self.gY = NULL
+
+    def __reduce__(self):
+        return self.__class__, (self.r, self.g)
+
+    def __dealloc__(self):
+        self._free()
+
+    cdef void _free(self) nogil:
+        if self.cost != NULL:
+            free(self.cost)
+            self.cost = NULL
+
+        if self.cost_prev != NULL:
+            free(self.cost_prev)
+            self.cost_prev = NULL
+        
+        if self.gX != NULL:
+            free(self.gX)
+            self.gX = NULL
+
+        if self.gY != NULL:
+            free(self.gY)
+            self.gY = NULL
+
+    cdef int reset(self, TSArray X) nogil:
+        self._free()
+        self.cost = <double*> malloc(sizeof(double) * X.shape[2])
+        self.cost_prev = <double*> malloc(sizeof(double) * X.shape[2])
+        self.gX = <double*> malloc(sizeof(double) * X.shape[2])
+        self.gY = <double*> malloc(sizeof(double) * X.shape[2])
+        if self.cost == NULL or self.cost_prev == NULL or self.gX == NULL or self.gY == NULL:
+            return -1
+
+    cdef double _distance(
+        self,
+        const double *s,
+        Py_ssize_t s_len,
+        double s_mean,
+        double s_std,
+        void *s_extra,
+        const double *x,
+        Py_ssize_t x_len,
+        Py_ssize_t *return_index=NULL,
+    ) nogil:
+        return erp_subsequence_distance(
+            s, 
+            s_len,
+            x,
+            x_len,
+            _compute_r(s_len, self.r),
+            self.g,
+            self.gX,
+            self.gY,
+            self.cost,
+            self.cost_prev,
+            return_index,
+        )
+
+    cdef Py_ssize_t _matches(
+        self,
+        double *s,
+        Py_ssize_t s_len,
+        double s_mean,
+        double s_std,
+        void *s_extra,
+        double *x,
+        Py_ssize_t x_len,
+        double threshold,
+        double **distances,
+        Py_ssize_t **indicies,
+    ) nogil:
+        return erp_subsequence_matches(
+            s,
+            s_len,
+            x,
+            x_len,
+            _compute_r(s_len, self.r),
+            self.g,
+            self.gX,
+            self.gY,
+            self.cost,
+            self.cost_prev,
             threshold,
             distances,
             indicies,